Vital fix for as-yet unused debug function.
Allowed pointer option to be set to "parent" or "all" (default; same as true) to specify whether just the parent element's or its and all its children's pointers should be set to cursor.  This works in Firefox and Chrome but IE has an issue and I have employed an odd workaround that I have not thoroughly tested.

Tried to facilitate multiple simultaneous text selection and if successful would have noted this among other things: Made it so when there's a child of a parent that's an input, it's text gets selected, too.  Added an option for this called select with settings "all" or "text" with all selecting the input including text and text just leaving the inputs be.  Also, a target can be more than one element now, etc.